% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_frequency.R
\name{flag_frequency}
\alias{flag_frequency}
\title{Flag aggregation by the frequency count method}
\usage{
flag_frequency(f)
}
\arguments{
\item{f}{A vector of flags containing the flags of a series for a given period.}
}
\value{
\code{flag_frequency} returns a character with a single character flag in case the highest frequency count
 is unique, or multiple character in case there are several flags with the highest frequency count.
}
\description{
Flag aggregation by the frequency count method
}
\examples{
flag_frequency(c("pe","b","p","p","u","e","d"))
flag_frequency(c("pe","b","p","p","eu","e","d"))


flags <- tidyr::spread(test_data[, c(1:3)], key = time, value = flags)
flag_frequency(flags[,5])
apply(flags[, c(2:ncol(flags))],2, flag_frequency)

}
