% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagr-package.R
\docType{package}
\name{flagr-package}
\alias{flagr}
\alias{flagr-package}
\title{flagr: What the Package Does (One Line, Title Case)}
\description{
For more information visit
    <https://2degrees-investing.org/>.
}
\author{
\strong{Maintainer}: Mauro Lepore \email{mauro@ixpantia.com} (0000-0002-1986-7988) [contractor]

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
