% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagr.R
\name{flagr}
\alias{flagr}
\title{Add a flag}
\usage{
flagr(flag, squared = FALSE, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{flag}{\href{https://www.iso.org/obp/ui/#search/code/}{ISO 3166-1-alpha-2} code of a country, see examples.}

\item{squared}{if \code{TRUE} will use square flag.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{id of \code{<span>} (optional).}
}
\description{
Add a flag icon
}
\details{
flag-icon css \href{http://flag-icon-css.lip.is/}{demo}.
}
\examples{
flagr("be") # belgian flag
flagr("gr") # greek flag
flagr("us", width = 100, height = 60) # us flag
flagr("ch", squared = TRUE) # swiss flag squared

}
\seealso{
flag-icons css \href{http://flag-icon-css.lip.is/}{demo}, on
\href{https://github.com/lipis/flag-icon-css}{github}
}
