% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-desc.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{This data set is a fictive data set with fictive values and flags for testing purposes.}
\format{A data frame with 195 rows and 4 variables:
\describe{
  \item{geo}{2 digit country code}
  \item{flags}{flag of the value}
  \item{time}{date of observation}
  \item{values}{value of the element}
}}
\source{
The source is in *.csv* format also available in the package.
}
\usage{
test_data
}
\description{
This data set is a fictive data set with fictive values and flags for testing purposes.
}
\keyword{datasets}
